added cool things
simplified stuff, made gun class(no array yet) and moved stuff from
projectile, shoot, and player classed to the gun class, also can do ROF,
armour piercing, latch, etc on guns. will add prototype thing to guns
like i did with  projectiles soon.

+++ SifiGame/SifiGame/js/projectile_collision.js
@@ -21,7 +21,7 @@
                     }
                     else {
                         if (platformcollection.array[platformcounter].type == 12) {
-                            platformcollection.array[platformcounter].health = platformcollection.array[platformcounter].health - playercollection.array[projectilecollection.array[projectilecounter].pindex].damage * projectilecollection.array[projectilecounter].damagemultiplier;
+                            platformcollection.array[platformcounter].health = platformcollection.array[platformcounter].health - projectilecollection.array[projectilecounter].damagemultiplier;
                             if (platformcollection.array[platformcounter].health == 0) {
                                 platformcollection.delete(platformcounter);
                             }
@@ -53,8 +53,8 @@
                         }
                         else
                         {
-                            playercollection.array[playercounter].health = playercollection.array[playercounter].health - playercollection.array[projectilecollection.array[projectilecounter].pindex].damage * projectilecollection.array[projectilecounter].damagemultiplier;
-                            game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].index, 1);
+                            playercollection.array[playercounter].health = playercollection.array[playercounter].health - projectilecollection.array[projectilecounter].damagemultiplier;
+                            game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].index, 10);
                             if ((Math.random() * 100) + 1 > 50) {
                                 var velmultiplyertestx = -1;
                             }
@@ -67,12 +67,13 @@
                             else {
                                 var velmultiplyertesty = 1;
                             }
-                            floating_numbercollection.add(playercollection.array[playercounter].x, playercollection.array[playercounter].y, ((Math.random() * 2) + 0.1) * velmultiplyertestx, ((Math.random() * 2) + 0.1) * velmultiplyertesty, playercollection.array[projectilecollection.array[projectilecounter].pindex].damage * projectilecollection.array[projectilecounter].damagemultiplier, 15, "rgb(200,75,75)")
+                            floating_numbercollection.add(playercollection.array[playercounter].x, playercollection.array[playercounter].y, ((Math.random() * 2) + 0.1) * velmultiplyertestx, ((Math.random() * 2) + 0.1) * velmultiplyertesty, projectilecollection.array[projectilecounter].damagemultiplier, 15, "rgb(200,75,75)")
                         }
 
-
-						projectilecollection.delete(projectilecounter);
-                        break;
+                        if (projectilecollection.array[projectilecounter].armourpiercing == 0) {
+                            projectilecollection.delete(projectilecounter);
+                        }
+						break;
 
                     }
